Исправлены {} скобки и сделано разбиение на модули v1
