+++ js/data.js
@@ -1,0 +1,29 @@
+const TYPE_VALUES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+
+const CHECK_VALUES = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURE_VALUES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const PHOTO_VALUES = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+export {TYPE_VALUES, CHECK_VALUES, FEATURE_VALUES, PHOTO_VALUES};
